requirements.txt
Cython==0.29.32
g2p-en==2.1.0
inflect==4.1.0
librosa==0.10.0.post2
matplotlib==3.5.3
pyworld==0.3.2
PyYAML==6.0
scikit-learn==1.1.2
scipy==1.9.1
soundfile==0.10.3.post1
tensorboard==2.12.0
tgt==1.4.4
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.46.1
unidecode==1.3.4
